Move RemoteTask.run into open4 mixin and add open3 mixin for jruby
