+++ Rakefile
@@ -14,9 +14,13 @@ Hoe.spec 'rake-remote_task' do
   self.rubyforge_name = 'hitsquad'
 
   dependency 'rake',     '~> 0.8'
-  dependency 'open4',    '~> 1.0'
+  dependency 'open4',    '~> 1.0' unless RUBY_PLATFORM =~ /java/
 
   multiruby_skip << "rubinius"
+
+  if RUBY_PLATFORM =~ /java/
+    spec_extras[ :platform ] = Gem::Platform.new( "java" )
+  end
 end
 
 # vim: syntax=ruby
